Relocated mocker.py, and added mock inside __init__.py
